Conflicts:
	OnlineCinemaProject/OnlineCinemaProject/Controllers/AccountController.cs
	OnlineCinemaProject/OnlineCinemaProject/Models/AccountViewModels.cs
	OnlineCinemaProject/OnlineCinemaProject/Models/IdentityModels.cs
	OnlineCinemaProject/OnlineCinemaProject/OnlineCinemaProject.csproj
	OnlineCinemaProject/OnlineCinemaProject/Views/Account/Index.cshtml
	OnlineCinemaProject/OnlineCinemaProject/Views/Account/Register.cshtml
	OnlineCinemaProject/OnlineCinemaProject/Web.config

+++ OnlineCinemaProject/OnlineCinemaProject/Models/subscription.cs
@@ -23,7 +23,7 @@ namespace OnlineCinemaProject.Models
         public int tariff_id { get; set; }
         public System.DateTime start { get; set; }
         public System.DateTime end { get; set; }
-        public double payment { get; set; }
+        public decimal payment { get; set; }
     
         public virtual ICollection<aspnetuser> aspnetusers { get; set; }
         public virtual tariff tariff { get; set; }
